### 2024-04-02 #6570 @Kevin-Mattheus-Moerman
Dear author, @danielasay , thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Before we proceed, please can you have a look at the following points:

- [ ] Your project appears to lack contributing guidelines at the moment (or these are at least not visible from the main repository page/README). Please work to add these and to mention/link to them in your README. You could for instance create a `CONTRIBUTING.md` file (see here for some examples: https://contributing.md/example/).
