+++ dwiqc/tasks/qsiprep_EQ.py
@@ -149,9 +149,9 @@ class Task(tasks.BaseTask):
 		-v
 		"""
 
-		#if os.path.isdir(f'{self._sub}_{self._ses}.qc'):
-		#	logging.warning('Output directory already exists. Removing and trying again.')
-		#	shutil.rmtree(f'{self._sub}_{self._ses}.qc')
+		if os.path.isdir(f'{self._sub}_{self._ses}.qc'):
+			logging.warning('Output directory already exists. Removing and trying again.')
+			shutil.rmtree(f'{self._sub}_{self._ses}.qc')
 
 		logging.info('Running eddy_quad...')
 		proc1 = subprocess.Popen(command, shell=True, stdout=subprocess.PIPE)
