+++ dwiqc/browser/__init__.py
@@ -10,11 +10,10 @@ home_dir = os.path.expanduser("~")
 
 
 def snapshot(url, saveto, container_dir):
-	proc1 = f'{container_dir}/chromium.sif --no-sandbox --headless --print-to-pdf={saveto} {url}'
-	output = subprocess.Popen(proc1, shell=True, stdout=subprocess.PIPE)
-	output.communicate()
+    proc1 = f'{container_dir}/chromium.sif --no-sandbox --headless --print-to-pdf={saveto} {url}'
+    output = subprocess.Popen(proc1, shell=True, stdout=subprocess.PIPE)
+    output.communicate()
     code = output.returncode
-
     if code == 0:
         logging.info('pdf conversion successful!')
     else:
