+++ dwiqc/tasks/qsiprep.py
@@ -1,0 +1,565 @@
+#### load necessary libraries
+
+import subprocess
+import os
+import yaml
+import logging
+from bids import BIDSLayout
+import sys
+import json
+import nibabel as nib
+import dwiqc.tasks as tasks
+import shutil
+from executors.models import Job
+import dwiqc.config as config
+import numpy as np
+from pprint import pprint
+import re
+import shutil
+from dipy.io import read_bvals_bvecs
+from dipy.core.gradients import gradient_table
+from datetime import datetime
+
+date = datetime.today().strftime('%Y-%m-%d')
+
+logger = logging.getLogger(__name__)
+
+
+class Task(tasks.BaseTask):
+	def __init__(self, sub, ses, run, bids, outdir, qsiprep_config, fs_license, container_dir=None, custom_eddy=False, no_gpu=False, output_resolution=None, tempdir=None, pipenv=None):
+		self._sub = sub
+		self._ses = ses
+		self._run = run
+		self._bids = bids
+		self._qsiprep_config = qsiprep_config
+		self._fs_license = fs_license
+		self._container_dir = container_dir
+		self._custom_eddy = custom_eddy
+		self._no_gpu = no_gpu
+		self._layout = BIDSLayout(bids)
+		self._output_resolution = output_resolution
+		super().__init__(outdir, tempdir, pipenv)
+
+
+	def check_container_path(self):
+		if self._container_dir:
+			try:
+				self._qsiprep_sif = f'{self._container_dir}/qsiprep.sif'
+			except FileNotFoundError:
+				logger.error(f'{self._container_dir}/qsiprep.sif does not exist. Verify the path and file name.')
+				sys.exit(1)
+			try:
+				self._fsl_sif = f'{self._container_dir}/fsl_6.0.4.sif'
+			except FileNotFoundError:
+				logger.error(f'{self._container_dir}/fsl_6.0.4.sif does not exist. Verify the path and file name.')
+				sys.exit(1)
+
+		else:
+			home_dir = os.path.expanduser("~")
+			self._qsiprep_sif = os.path.join(home_dir, '.config/dwiqc/containers/qsiprep.sif')
+			self._fsl_sif = os.path.join(home_dir, '.config/dwiqc/containers/fsl_6.0.4.sif')
+
+
+	def calc_mporder(self):
+		## if --no-gpu argument is passed, make mporder equal to 1
+		if self._no_gpu:
+			mporder = 1
+			return mporder
+
+		# need to get the length of the SliceTiming File (i.e. number of slices) divided by the "MultibandAccelerationFactor". From there, 
+		# divide that number by 3. That's the number passed to mporder.
+
+		# this will grab the dwi file, pop it off the list, get the slice timing metadata, then grab the length of the slice timing array
+		num_slices = len(self._layout.get(subject=self._sub, session=self._ses, run=self._run, suffix='dwi', extension='.nii.gz').pop().get_metadata()['SliceTiming'])
+
+		multiband_factor = self._layout.get(subject=self._sub, session=self._ses, run=self._run, suffix='dwi', extension='.nii.gz').pop().get_metadata()['MultibandAccelerationFactor']
+
+		mporder = (num_slices / multiband_factor) // 3
+
+		return int(mporder)
+
+
+
+	# this method checks if the user has passed a desired output resolution to dwiqc.py
+	# if they haven't, the resolution of the T1w input image will be used.
+
+	def check_output_resolution(self):
+		if not self._output_resolution:
[…]
