### 2024-04-09 9a6910f
add two more files to xnat-artifacts
