+++ dwiqc/tasks/prequal.py
@@ -1,674 +1,0 @@
-#### load necessary libraries
-import tempfile
-import subprocess
-import yaml
-import os
-import logging
-from bids import BIDSLayout
-import sys
-import json
-import dwiqc.tasks as tasks
-import shutil
-from executors.models import Job
-import dwiqc.config as config
-import numpy as np
-from pprint import pprint
-import re
-import nibabel as nib
-from datetime import datetime
-
-
-logger = logging.getLogger(__name__)
-
-
-# pull in some parameters from the BaseTask class in the __init__.py directory
-
-class Task(tasks.BaseTask):
-	def __init__(self, sub, ses, run, bids, outdir, prequal_config, fs_license, container_dir=None, no_gpu=False, tempdir=None, pipenv=None):
-		self._sub = sub
-		self._ses = ses
-		self._run = run
-		self._bids = bids
-		self._prequal_config = prequal_config
-		self._fs_license = fs_license
-		self._container_dir = container_dir
-		self._no_gpu = no_gpu
-		self._layout = BIDSLayout(bids)
-		self._date = datetime.today().strftime('%Y-%m-%d')
-		super().__init__(outdir, tempdir, pipenv)
-
-
-	# create an INPUTS dir next to the OUTPUTS dir
-	def copy_inputs(self, inputs_dir):
-		"""
-		try:
-			os.makedirs(inputs_dir)
-		except FileExistsError:
-			pass
-		"""
-
-		os.makedirs(inputs_dir, exist_ok=True)
-		
-		all_files = self._layout.get(subject=self._sub, session=self._ses, return_type='filename') # get a list of all of the subject's files
-
-		# copy the all the subject's files into the INPUTS directory
-		for file in all_files:
-			basename = os.path.basename(file)
-			dest = os.path.join(inputs_dir, basename)
-			shutil.copy(file, dest)
-
-		self.create_bfiles(inputs_dir)
-		
-
-	# the fieldmap data needs accompanying 'dummy' bval and bvec files that consist of 0's
-	def create_bfiles(self, inputs_dir):
-		# get a list of all the fmap files that end with .json (this it's helpful to have a file with just one extension)
-
-		fmap_files = self._layout.get(subject=self._sub, session=self._ses, suffix='epi', extension='.nii.gz', return_type='filename')
-
-		if not fmap_files:
-			self._layout.get(subject=self._sub, session=self._ses, suffix='epi', extension='.nii', return_type='filename')
-
-		# get the basename of the file and then remove the extension
-		for fmap in fmap_files:
-
-			basename = os.path.basename(fmap)
-
-			no_ext = basename.split('.', 1)[0]
-
-			# get the number of volumes in the data file
-			try:
-				num_vols = nib.load(fmap).shape[3]
-			except IndexError:
-				num_vols = 1
-
-			# create a .bval file same number of rows of 0 as there are volumes
-
-			with open(f'{inputs_dir}/{no_ext}.bval', 'w') as bval:
-				rows_written = 0
-				while rows_written < num_vols:
-					if rows_written == 0:
-						bval.write('0')
-					elif rows_written == num_vols - 1:
-						bval.write(' 0\n')
-					else:
-						bval.write(' 0')
-					rows_written += 1
-
-			# create .bvec file with 3 row 0's equal to the number of volumes
-
-			with open(f'{inputs_dir}/{no_ext}.bvec', 'w') as bvec:
-				row_to_write = ''
-				# write the same number of 0's as there are volumes
-				for _ in range(num_vols):
-					row_to_write += '0 '
-				# remove any trailing whitespace
-				row_to_write = row_to_write.strip()
-
-				# add three rows to bvec file
[…]
