### 2024-04-22 1807dd1
add --work-dir argument for prequal
