### 2024-04-23 e45c7c6
bind in bids dir when singularity is called
