+++ dwiqc/tasks/prequal.py
@@ -543,7 +543,7 @@ class Task(tasks.BaseTask):
 			return None
 	def bind_environmentals(self):
 	
-		bind = [self._outdir, self._tempdir, self._fs_license]
+		bind = [self._bids, self._tempdir, self._fs_license]
 		
 		os.environ["SINGULARITY_BIND"] = ','.join(bind)
 
