+++ dwiqc/tasks/qsiprep.py
@@ -486,7 +486,7 @@ class Task(tasks.BaseTask):
 
 	def bind_environmentals(self):
 	
-		bind = [self._outdir, self._tempdir, self._fs_license]
+		bind = [self._bids, self._tempdir, self._fs_license]
 		
 		os.environ["SINGULARITY_BIND"] = ','.join(bind)
 
