### 2024-04-23 da16d72
add exist_ok=True to os.makedirs commands
