### 2024-04-23 c9abd4c
update tempdir/workdir configuration. add logging statements
