### 2024-04-23 959e9d5
get rid of unncecessary print statement
