### 2024-04-25 8ddeb3d
replace `os.chdir()` with `singularity exec --pwd` within extract_b0_vol, add more logging, and validate inputs and outputs
