### 2024-04-25 f9b040c
add more logging and create error class for browser __init__.py
