### 2024-04-25 afc5f33
more edits to browser __init__.py
