+++ .github/workflows/pylint.yml
@@ -20,4 +20,4 @@ jobs:
         pip install pylint
     - name: Analysing the code with pylint
       run: |
-        pylint $(git ls-files '*.py') --output=lint.txt || true
+        pylint $(git ls-files '*.py') || true
