+++ .github/workflows/pylint.yml
@@ -20,4 +20,10 @@ jobs:
         pip install pylint
     - name: Analysing the code with pylint
       run: |
-        pylint $(git ls-files '*.py') || true
+        pylint $(git ls-files '*.py') --output=lint.txt || true
+
+    - name: Upload Artifact
+      uses: actions/upload-artifact@v3
+      with:
+        name: lint.txt
+        path: lint.txt
