### 2024-04-29 d8aa01c
add path to pylint workflow
