### 2024-04-29 6158145
updates to browser module
