+++ dwiqc/cli/get.py
@@ -1,7 +1,6 @@
-print(os.getcwd())
-
 import os
 import re
+sys.exit()
 import sys
 import json
 import yaml
