### 2024-04-29 863e1c4
fail pylint on error
