### 2024-04-29 dd34fb0
fail pylint on error
