+++ .github/workflows/pylint.yml
@@ -18,6 +18,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install pylint
+        pip install .
     - name: Analysing the code with pylint
       run: |
-        pylint $(git ls-files '*.py') --fail-on E || true
+        pylint $(git ls-files '*.py') --fail-on E
