### 2024-04-29 a10c906
fail pylint on error
