+++ .github/workflows/pylint.yml
@@ -21,4 +21,4 @@ jobs:
         pip install .
     - name: Analysing the code with pylint
       run: |
-        pylint $(git ls-files '*.py') --fail-on E
+        pylint $(git ls-files '*.py') || exit $(($? & 3))
