### 2024-04-29 5d360b4
add more checking and logging around b0 volume detection and extraction in qsiprep.py
