+++ dwiqc/tasks/qsiprep.py
@@ -360,7 +360,13 @@ class Task(tasks.BaseTask):
 
 		list_of_bvals = []
 
+		logger.info('running dipy.core.gradients.gradient_table')
 		gtab = gradient_table(bvals, bvecs)
+
+		bval_min = min(gtab.bvals)
+		logger.info(f'minimum bval is {bval_min}')
+		if bval_min > 50:
+			logger.warning(f'minimum bval is greater than 50')
 
 		for val in gtab.bvals:
 			list_of_bvals.append(str(val))
@@ -368,23 +374,30 @@ class Task(tasks.BaseTask):
 		b0_volumes = []
 
 		for idx, bval in enumerate(list_of_bvals):
-			if bval == '0.0':
+			if bval == str(bval_min):
+				logger.info(f' found b0 volume at index {idx} ({bval} == {bval_min})')
 				b0_volumes.append(str(idx))
 
 		b0_string = ','.join(b0_volumes)
 
+		if not b0_string.strip():
+			raise Exception(f'failed to find any b0 volumes in {list_of_bvals}')
 
 		os.makedirs(self._outdir, exist_ok=True)
 
 		try:
-
-			extract_command = f"singularity exec /{self._fsl_sif} /APPS/fsl/bin/fslselectvols -i {dwi_full_path} -o {epi_out_path} --vols={b0_string}"
-
+			extract_command = f'singularity exec /{self._fsl_sif} /APPS/fsl/bin/fslselectvols -i {dwi_full_path} -o {epi_out_path} --vols={b0_string}'
+			logger.info(f'executing {extract_command}')
+			bindings = os.environ.get('SINGULARITY_BIND', None)
+			logger.info(f'SINGULARITY_BIND environment variable contains "{bindings}"')
 			proc1 = subprocess.check_output(extract_command, shell=True, stderr=subprocess.STDOUT, text=True)
-
+			if not os.path.exists(epi_out_path):
+				logger.critical(f'fslselectvols failed to produce "{epi_out_path}"')
+				raise subprocess.CalledProcessError(returncode=1, cmd=extract_command)
+			logger.info(f'found fslselectvols derived file "{epi_out_path}"')
 		except subprocess.CalledProcessError as e:
 			print(e.stdout)
-
+			raise e
 
 	def insert_json_value(self, key, value, json_file):
 		"""
