+++ dwiqc/cli/get.py
@@ -1,3 +1,4 @@
+print(os.getcwd())
 import os
 import re
 import sys
