### 2024-05-01 a4327d9
updates to xnat installation instructions, now includes nginx config
