+++ docs/source/developers.rst
@@ -75,6 +75,135 @@ Once the plugin has been compiled, move the resulting ``.jar`` into your XNAT pl
 
     mv ./build/libs/dwiqc-plugin-1.0.0.jar ${XNAT_HOME}/plugins/
 
+Configuring Nginx
+^^^^^^^^^^^^^^^^^
+You will have to do a little bit of extra configuration in the ``nginx.conf`` file on the XNAT server. Without this configuration, XNAT will start downloading all of the DWIQC reports to your local machine and you won't be able to view them on the browser. Here's a way around that!
+
+First, navigate to your ``nginx`` directory. This might be located at ``/etc/nginx`` if you're running XNAT on "bare metal" or under your ``xnat-docker-compose`` directory if you're using that to run your XNAT. Once you're there, cd into the ``conf.d`` directory and open up the server config file using your favorite text editor. ``xnat-docker-compose`` users may have a file named ``nginx.conf``. 
+
+Your ``nginx.conf`` file may look something like this:
+
+.. code-block:: yaml
+
+
+    #user www-data;
+    worker_processes auto;
+    #pid /run/nginx.pid;
+    events {
+      worker_connections 1024;
+      # multi_accept on;
+    }
+    http {
+      ##
+      # Basic Settings
+      ##
+      sendfile on;
+      tcp_nopush on;
+      tcp_nodelay on;
+      keepalive_timeout 100;
+      types_hash_max_size 2048;
+      # server_tokens off;
+      # server_names_hash_bucket_size 64;
+      # server_name_in_redirect off;
+      include /etc/nginx/mime.types;
+      default_type application/octet-stream;
+      ##
+      # SSL Settings
+      ##
+      #ssl_protocols TLSv1 TLSv1.1 TLSv1.2; # Dropping SSLv3, ref: POODLE
+      #ssl_prefer_server_ciphers on;
+      ##
+      # Logging Settings
+      ##
+      access_log /var/log/nginx/access.log;
+      error_log /var/log/nginx/error.log;
+      ##
+      # Gzip Settings
+      ##
+      gzip on;
+      gzip_disable "msie6";
+      # gzip_vary on;
+      # gzip_proxied any;
+      # gzip_comp_level 6;
+      # gzip_buffers 16 8k;
+      # gzip_http_version 1.1;
+      # gzip_types text/plain text/css application/json application/javascript text/xml application/xml application/xml+rss text/javascript;
+      ##
+      # Virtual Host Configs
+      ##
+      #include /etc/nginx/conf.d/*.conf;
+      #include /etc/nginx/sites-enabled/*;
+      server {
+        listen 80;
+        return                301 https://cbscentral02.rc.fas.harvard.edu$request_uri;
+      } 
+
+      server {
+        listen 443 ssl;
+        ssl_certificate       /etc/pki/tls/cbscentral02_rc_fas_harvard_edu.crt;
+        ssl_certificate_key   /etc/pki/tls/cbscentral02_rc_fas_harvard_edu.key;
+        location / {
+            proxy_pass                          http://xnat-web:8080;
+            proxy_redirect                      http://xnat-web:8080 $scheme://localhost;
+            proxy_set_header Host               $host;
+            proxy_set_header X-Real-IP          $remote_addr;
+            proxy_set_header X-Forwarded-Host   $host;
+            proxy_set_header X-Forwarded-Server $host;
+            proxy_set_header X-Forwarded-For    $proxy_add_x_forwarded_for;
+            proxy_connect_timeout               150;
+            proxy_send_timeout                  100;
+            proxy_read_timeout                  100;
+            proxy_buffers                       8 4k;
+            proxy_max_temp_file_size            2048m;
+            client_max_body_size                0;
+            client_body_buffer_size             128k;
[…]
