### 2024-05-01 3525d99
further clarification and text highlighting in nginx config file section
