### 2024-05-20 c42c314
add try except logic for getting metadata
