### 2024-06-25 235fde4
add beginnings of truncation for fmaps
