### 2024-06-25 d363cec
print out fslselectvols command
