### 2024-06-25 0e2bfff
testing b0 creation mehtod
