+++ dwiqc/tasks/prequal.py
@@ -151,7 +151,7 @@ class Task(tasks.BaseTask):
 				logger.critical(f'fslselectvols command failed')
 				raise subprocess.CalledProcessError(returncode=proc.returncode, cmd=cmdline)
 
-			truncated_fmaps.append(fmap_dir + f'/fmap_basename')
+			truncated_fmaps.append(fmap_dir + f'/{fmap_basename}')
 
 
 		print(truncated_fmaps)
