### 2024-06-25 8e570d0
ready for testing on adni data
