### 2024-06-25 67cd1ee
remove final testing code
