### 2024-06-25 8306a87
new version with fmap b0volume extraction for prequal
