### 2024-06-26 d5bc5a4
update that copies b0volume fmaps to inputs dir for prequal
