### 2024-06-26 7c0dc13
modify --pwd arg in b0volume fmap creation
