### 2024-06-26 7379076
new version with truncation happening in copy_inputs method
