### 2024-06-27 #6570 @spinicist
Hello, I know a few diffusion experts who might be interested in this but I don't know if they are in the JOSS database already. If you drop me an e-mail (my work one is easy to find, look at the READMEs for my main repos) then I can provide contact details.
