### 2024-06-27 dab826d
fix typo in fslroi command
