### 2024-07-02 fbd9b3c
fix the truncation so that all fmap files get truncated
