+++ dwiqc/tasks/qsiprep.py
@@ -355,11 +355,11 @@ class Task(tasks.BaseTask):
 
 
 	def execute_fmap_truncation(self):
-	"""
-	If truncate fmaps is true, this method will go through all of the fmaps in the fmap BIDS dir and make sure they are only one volume in length
-	"""
+		"""
+		If truncate fmaps is true, this method will go through all of the fmaps in the fmap BIDS dir and make sure they are only one volume in length
+		"""
 		if self._truncate_fmap:
-			fmap_files = self._layout.get(subject=self._sub, session=self._ses, suffix='epi', extension='.nii.gz', return_type='filename')	
+			fmap_files = self._layout.get(subject=self._sub, session=self._ses, suffix='epi', extension='.nii.gz', return_type='filename')
 
 			for fmap in fmap_files:
 				shape = nib.load(fmap).get_fdata().shape
