### 2024-07-02 bedb566
add logging
