### 2024-07-02 ef631e1
reload layout so all fmaps get read in correctly
