### 2024-07-21 #6974 @cjohnevans
Hi @danielasay @samhforbes

Thanks for the opportunity to review, however, I don’t think I’m able to be able to perform an adequate review at this time.  The package looks very useful, and certainly it would be keen to get this implemented at our site, however we are quite a long way short of the prerequisites to be able to do a sufficient review at this time.  We do manage our data with an XNAT server, but we make limited use of XNAT scripts, similarly, we have partially implemented the BIDS data structure, but it isn’t our standard method of data handling yet.

These are shortcomings of our site, rather than the software, but I suspect this means we’re a few months of implementation time away from being able to do a proper test of the code.

If it’s useful, I have a few comments, based on my reading of the user documentation and installation in lieu of a proper review.  These comments are based on https://dwiqc.readthedocs.io/en/latest/ as of 21/7/24.

-	Prerequisites should be clearly defined at the start of the User Documentation (e.g. XNAT version, GPU node specs, prequal version, qsiprep version, …)
-	User Documentation reads as if it’s targeted at local users at the author's institution with a pre-installed instance rather than new, external users.
-	Similarly, it would make more sense to me to begin with the ‘outside XNAT’ mode in the documentation.  The non-XNAT implementation is defined rather late on in the documentation (https://dwiqc.readthedocs.io/en/latest/xnat.html#process-mode )
-	Given the necessary pre-requisites, it would be good to package some example data in the appropriate BIDS format for the validation of installation.  While this may be challenging for data protection or data size reasons, possible a suitable open-access dataset could be identified for testing.
-	dwiqc installation worked as expected.

Apologies again that I won’t be able to participant further at this stage, but I will keep an eye on this repo and work towards getting an implementation at our site in the future.
