### 2024-07-31 864bde6
add container directory check for chromium
