### 2024-07-31 ece59ba
add default value for container dir in snapshot function
