### 2024-08-07 d2ed270
add slurm job id as directory name under workdir
