### 2024-08-07 5917aee
add logic to copy input bvec if eddy quad fails on qsiprep output
