### 2024-08-07 9b9974d
change positioning of slurm job id in tempdir
