### 2024-08-08 91f0e49
add slurm job to copying commands in qsiprep_EQ
