+++ dwiqc/tasks/qsiprep_EQ.py
@@ -44,9 +44,9 @@ class Task(tasks.BaseTask):
 
 		# Define working directory for the subject
 
-		dwi_preproc_string = self.match_preproc_string(f'{self._tempdir}/qsiprep_{date}/SLURM_JOB_{os.environ['SLURM_JOB_ID']}/{self._ses}/qsiprep_wf/single_subject_{self._sub}_wf')
+		dwi_preproc_string = self.match_preproc_string(f"{self._tempdir}/qsiprep_{date}/SLURM_JOB_{os.environ['SLURM_JOB_ID']}/{self._ses}/qsiprep_wf/single_subject_{self._sub}_wf")
 
-		qsiprep_work_dir = f'{self._tempdir}/qsiprep_{date}/SLURM_JOB_{os.environ['SLURM_JOB_ID']}/{self._ses}/qsiprep_wf/single_subject_{self._sub}_wf/{dwi_preproc_string}/hmc_sdc_wf'
+		qsiprep_work_dir = f"{self._tempdir}/qsiprep_{date}/SLURM_JOB_{os.environ['SLURM_JOB_ID']}/{self._ses}/qsiprep_wf/single_subject_{self._sub}_wf/{dwi_preproc_string}/hmc_sdc_wf"
 
 		# Define eddy quad destination directory
 
