### 2024-08-08 12c95f4
add logic to prequal_EQ to run eddy quad on spit output
