+++ dwiqc/tasks/prequal_EQ.py
@@ -43,22 +43,12 @@ class Task(tasks.BaseTask):
 		# define log file
 		logging.basicConfig(filename=f"{self._outdir}/logs/dwiqc-prequal.log", encoding='utf-8', level=logging.DEBUG)
 
-		# copy the necessary file to EDDY directory
-
-		logging.info('copying dwmri.nii.gz file to EDDY dir')
-		if os.path.isfile(f'{self._outdir}/PREPROCESSED/dwmri.nii.gz'):
-			shutil.copy(f'{self._outdir}/PREPROCESSED/dwmri.nii.gz', f'{self._outdir}/EDDY/{self._sub}_{self._ses}.nii.gz')
-
-		else:
-			logging.error('PREPROCESSED/dwmri.nii.gz doesn\'t exist, which likely means prequal failed. exiting.')
-			sys.exit()
-
 		# grab name of slspec file
 		logging.info(f'grabbing slspec file from {self._outdir}')
 		try:
 			for file in os.listdir(self._outdir):
 				if 'slspec' in file and file.endswith('.txt'):
-					spec_file = file
+					self._spec_file = file
 		except FileNotFoundError:
 			logging.error('spec file not found. exiting.')
 			sys.exit()
@@ -72,24 +62,55 @@ class Task(tasks.BaseTask):
 				os.rename(file, new_name)
 
 
+		self.check_split_preproc()
+
+	def check_split_preproc(self):
+		# copy the necessary file to EDDY directory
+
+		all_files = [file for file in os.listdir(f'{self._outdir}/PREPROCESSED')]
+		if 'dwmri.nii.gz' in all_files:
+			logging.info('only one preproc file found, running eddy quad')
+			self.run_eddy_quad('dwmri.bval', 'dwmri.bvec')
+			self.extract_b0_vol_regular()
+
+		else:
+			logging.warning('PREPROCESSED/dwmri.nii.gz doesn\'t exist, checking for split output')
+			main_scans = [file for file in all_files if '_dwi_' in file and file.endswith('preproc.nii.gz')]
+			logger.info(f'found these main diffusion scans {main_scans}')
+			self.split_output_processing(main_scans)
+
+	def split_output_processing(self, scans):
+		if scans:
+			logger.info('split scans found')
+
+			for idx,scan in enumerate(scans, 1):
+				self.run_eddy_quad(scan.replace('.nii.gz', '.bval'), scan.replace('.nii.gz', '.bvec'), idx)
+				self.extract_b0_vol_split(scans, idx)
+
+		else:
+			logger.error('split scan output not found, exiting')
+			sys.exit()
+
+
+	def run_eddy_quad(self, bval, bvec, run=1):
 
 		eddy_quad = f"""singularity exec \
 		--pwd {self._outdir}/EDDY \
 		{self._fsl_sif} \
 		/APPS/fsl/bin/eddy_quad \
-		{self._sub}_{self._ses} \
+		{self._sub}_{self._ses}_{run} \
 		-idx index.txt \
 		-par acqparams.txt \
 		--mask=eddy_mask.nii.gz \
-		--bvals={self._outdir}/PREPROCESSED/dwmri.bval \
-		--bvecs={self._outdir}/PREPROCESSED/dwmri.bvec \
+		--bvals={self._outdir}/PREPROCESSED/{bval} \
+		--bvecs={self._outdir}/PREPROCESSED/{bvec} \
 		--field {self._outdir}/TOPUP/topup_field.nii.gz \
-		-s {self._outdir}/{spec_file} \
+		-s {self._outdir}/{self._spec_file} \
 		-v"""
 
-		if os.path.isdir(f'{self._sub}_{self._ses}.qc'):
[…]
