### 2024-08-08 1352e3c
add rename_eddy_files method to make sure file names align with run number
