### 2024-08-08 8f3a4fd
make sure correct nii file is copied over to eddy dir
