### 2024-08-08 fe5b2d7
add function that edits index file to match number of volumes in split output
