### 2024-08-09 71365ed
turn off split processing for EQ
