### 2024-08-12 41cf2ec
remove run referecnes in prequal eddy quad
