### 2024-08-13 1b75828
add logic to catch jobs that do not have a slurm job id
