### 2024-08-13 #6974 @mwegrzyn
Hi @samhforbes ,
thank you for your message. I am done with my review and have discussed all open issues with the author (https://github.com/harvard-nrg/dwiqc/issues?q=is%3Aissue+is%3Aclosed).
Best,
Martin
