### 2024-08-13 d344d33
add in support to customize ol_std parameter for eddy in prequal
