### 2024-08-13 44953ab
fix string logic
