### 2024-08-14 fa43737
add in slurm job id to process script and propogate from there
