### 2024-08-14 8ba7ed9
use self._slurm_job_id
