### 2024-08-14 399a4f7
add SLURM_JOB_ prefix to number
