### 2024-08-14 8dc787e
add in mem to tandem mode
