### 2024-09-05 5247754
add logging to json reading
