### 2024-09-05 e365236
debug logging
