### 2024-09-05 4a45510
add logging
