### 2024-09-05 b76f1d6
add new method to filter json files
