### 2024-09-06 #6974 @samhforbes
Hi @jelleveraart you can ask editorialbot to generate your checklist for you, and then start your review. Typically we address issues in the software repository directly, and then link back to here. Let me know if I can help with anything!
