### 2024-09-09 3d6f577
making default resolution 1.0 if t1 not provided
