### 2024-09-10 a9996b7
make self._run a string in qsiprep_EQ
