### 2024-09-11 20497f9
use path objects for copying back up bval and bvec files
