### 2024-09-23 dadb16e
cut off prefix for bval copy
