### 2024-10-01 438a262
full path to bval file
