### 2024-10-02 614aa79
add logic to rename copied bvec file
