### 2024-10-07 ebd5ece
add try except block for calculating mporder
