### 2024-10-07 c6b4a68
verify bval size
