### 2024-10-07 b2a2994
fix variable error
