### 2024-10-08 be0e078
expand match bval method
