+++ dwiqc/tasks/qsiprep_EQ.py
@@ -115,7 +115,12 @@ class Task(tasks.BaseTask):
 	def match_bval(self, dir_path):
 		for file in os.listdir(dir_path):
 			if file.endswith('.bval') and str(self._run) in file:
-				return file
+				bval = file
+		if bval is None:
+			for file in os.listdir(dir_path):
+				if file.endswith('.bval'):
+					bval = file
+		return bval
 
 
 	def match_preproc_string(self, input_dir):
