### 2024-10-10 d37a28b
fix bval assignment
