### 2024-10-11 decaaa5
add logic to correct bvals in event of bval bvec mis-match
