### 2024-10-23 #6974 @jelleveraart
Hi @danielasay

Please find some comments related to the software paper.  There was a misunderstanding on my end that led to the delay in my review, but I am currently prioritizing this assignment.  Comments on functionality and documentation will follow later this week.

**A statement of need:**
The goal of the DWIQC is not defined unambiguously.  Whereas the user documents clearly specifies that DWIQC is a diffusion MRI quality control pipeline, the manuscript indicates that DWIQC performs image preprocessing and quality control. I do understand that preprocessing of data is necessary to perform the quality control, but it results in lack of clarity regarding the output.   Does DWIQC ultimately provides access to the endpoints of diffusion analyses (e.g. parametric maps or connectivity matrices) that are generated by PREQual and QSIprep? It would be of interest to clearly specify the outputs in the the manuscript and the user documentation.
In “process: Expected Output”, you provide an option to download examples of quality reports, but the download links are not active. In “Understanding the report page”, there is more information given, but only for users uploading DWIQC output to an XNAT instance. Therefore it remains unclear what output is to be expected for someone who runs DWIQC on HPC or local systems

**State of the field:**
It is not well defined in the paper or documentation how DWIQC compares to the direct use of containerized versions of QSIprep or PREqual, when run on HPC or local systems,  without leveraging the XNAT integration.

**References:**
Recent developments such as NeuroDesk  (https://doi.org/10.1038/s41592-023-02145-x) and BrainLife (https://doi.org/10.1038/s41592-024-02237-2) are relevant to discuss and reference in context of this paper.
