### 2024-10-24 #6974 @jelleveraart
Hi @danielasay (and @samhforbes )

Please find some comments and questions related to the Documentation:

**General remarks:**
- In the documentation, dwiqc is primarily presented as a user interface for the XNAT informatics and data management platform. When not using XNAT, users might struggle to navigate the user guidelines and, for those users, the added value of dwiqc compared to QSIprep and PreQual remains unclear to me.
- A clear separation between XNAT and non-XNAT users would be beneficial. Non-XNAT users would definitely benefit from a section on data requirements and organization. For example, the introduction of fieldmaps and reverse polarity scans are given in the ‘get” section that  non-xnat users are suggested to skip.
- PreQual requires the creation of configuration files. How are those files generated from the BIDS data for non-XNAT users and has the tool shown to be robust across vendors? If not, please specify which vendors have been tested.
- Understanding the Report Page is stated to only be relevant to XNAT users. Yet, it would be important to help non-XNAT users navigate the QC reports of dwiqc. A reference  to user’s guides of QSIprep and PREqual might be sufficient, but a dedicated section in the User Documentation would be appropriate.

**Process: Expected Output:**
- The reported processing times for PreQual and QSIprep are excessively long under the assumption that GPUs can be leveraged. Do QSIprep and PreQual benefit from the required GPUs? If so, are there requirements in terms of cuda version to ensure compatibility?
- “Download an example here” is not working
- Are the outputs of QSIprep and PREQual saved in their entirety? If yes, then the users could access the endpoints of diffusion analysis without running QSIprep or PreQual again. In that case, a reference that describes all output files of QSIprep or Prequal would be of interest to the user. Currently, the user guidelines are limited to the QC reports.

**Process: Advanced Usage:**
- -qsiprep-config and –prequal-config require config (*.yaml) arguments. An example is provided, but information on how to customize the example is missing. A list of the arguments listed in the config files and the possible choices would be relevant, but a simple reference to user guidelines of QSIprep and Prequal might be sufficient here.
- -artifacts-dir is only  introduced in the summary table
