### 2024-11-13 #6974 @samhforbes
@jelleveraart can I check since you've ticked everything off, have the documentation issues you've flagged here been addressed satisfactorily?
