+++ paper/paper.md
@@ -32,12 +32,12 @@ bibliography: citations.bib
 
 # Summary
 
-Diffusion weighted neuroimaging is an MR modality that maps white-matter microstructure in the human brain. `DWIQC` serves as a robust quality assurance preprocessing tool for diffusion weighted images as well as a means to facilitate data management and sharing via the XNAT platform. `DWIQC` utilizes analysis tools developed by FSL [@smith_advances_2004], Prequal [@cai_prequal:_2021], Qsiprep [@cieslak_qsiprep:_2021] and MRtrix [@tournier_mrtrix:_2012] to perform first level preprocessing of diffusion weighted images and to assess data quality through quantitative metrics. `DWIQC` utilizes containerized versions of the aforementioned software to ensure reproducibility of results and portability. `DWIQC` generates an aggregated report of summary metrics and images from the output of analysis software, which can be uploaded to the XNAT [@marcus_extensible_2007] data management platform, though upload to XNAT is not necessary.
+Diffusion weighted neuroimaging is an MR modality that maps white-matter microstructure in the human brain. `DWIQC` serves as a robust quality assurance preprocessing tool for diffusion weighted images as well as a means to facilitate data management and sharing via the XNAT platform. `DWIQC` utilizes analysis tools developed by FSL [@smith_advances_2004], Prequal [@cai_prequal:_2021], Qsiprep [@cieslak_qsiprep:_2021] and MRtrix [@tournier_mrtrix:_2012] to perform first level preprocessing of diffusion weighted images and to assess data quality through quantitative metrics. `DWIQC` utilizes containerized versions of the aforementioned software to ensure reproducibility of results and portability. `DWIQC` generates an aggregated report of summary metrics and images from the output of analysis software, including parametric maps and connectivity matrices from Prequal and Qsiprep, which can be uploaded to the XNAT [@marcus_extensible_2007] data management platform, though uploading to XNAT is not necessary. All outputs from software used are stored in the user specified output directory.
 
 
 # Statement of need
 
-Diffusion neuroimaging is a burgeoning field with huge potential to deepen our understanding of the brain. While exciting, it also means that acquisition parameters, study designs, and theoretical analysis frameworks vary greatly. `DWIQC` is an effort to make diffusion imaging analysis and quality assurance accessible to researchers with diverse experimental designs. By leveraging the respective strengths of various diffusion imaging analysis tools, `DWIQC` appeals to a broad range of users who employ different data acquisition approaches. Additionally, large, multi-site studies can can use the standardized preprocessing to ensure uniformity in data preprocessing, data quality metrics and downstream analysis.
+Diffusion neuroimaging is a burgeoning field with huge potential to deepen our understanding of the brain. While exciting, it also means that acquisition parameters, study designs, and theoretical analysis frameworks vary greatly. `DWIQC` is an effort to make diffusion imaging analysis and quality assurance accessible to researchers with diverse experimental designs. By leveraging the respective strengths of various diffusion imaging analysis tools, `DWIQC` appeals to a broad range of users who employ different data acquisition approaches. Additionally, large, multi-site studies can can use the standardized preprocessing to ensure uniformity in data preprocessing, data quality metrics and downstream analysis. `DWIQC` serves as a one stop destination for users who want a seemless integration of Prequal, Qsiprep and FSL Eddy Quad with the flexibility of catering to diverse acquisition protocols. Users can configure settings in `DWIQC` that will be applied across all the software that is used, abstracting away the sometimes convoluted task of doing so for each software pipeline individually.
 
[…]
