+++ docs/source/xnat.rst
@@ -67,7 +67,7 @@ install-containers, get, process and tandem modes
 get mode
 ^^^^^^^^
 .. note::
-    *get* mode is only applicable if you have an XNAT instance (such as CBS Central) you're going to interact with. Please feel free to skip to the `process <#process-mode>`_ mode section if you're only going to use *DWIQC* outside of XNAT.
+    Some sections of *get* mode are only applicable if you have an XNAT instance you're going to interact with. However, non-XNAT users may benefit from reviewing the section on fieldmaps and experimental design (immediately below). Please feel free to skip to the `process <#process-mode>`_ mode section if you're only going to use *DWIQC* outside of XNAT and already have familiarity with diffusion imaging experimental design.
 
 get: Overview
 """""""""""""
@@ -352,7 +352,9 @@ Command Example:
 process: Expected Output
 """"""""""""""""""""""""
 
-*DWIQC* runtime varies based on available resources, size of data and desired processing steps. Users should expect one session to take 3-5 hours to complete prequal and 7-10 hours to complete qsiprep. Prequal and qsiprep are run in parallel, so total processing time rarely exceeds 10 hours. *DWIQC* also makes use of the FSL tool eddy quad. Eddy quad runs a series of quality assesment commands to generate images and quantitative metric tables. Eddy quad doesn't take more than 10 minutes to run in most cases. A successful *DWIQC* run will contain output from all three of these software packages. 
+*DWIQC* runtime varies based on available resources, size of data and desired processing steps. On an HPC with gpu nodes, users should expect one session to take 1-2 hours to complete prequal and 4-6 hours to complete qsiprep. Prequal and qsiprep are run in parallel, so total processing time rarely exceeds 8-10 hours. *DWIQC* also makes use of the FSL tool eddy quad. Eddy quad runs a series of quality assesment commands to generate images and quantitative metric tables. Eddy quad doesn't take more than 10 minutes to run in most cases. A successful *DWIQC* run will contain output from all three of these software packages. It is worth noting that *DWIQC* retains all generated output files from Prequal and Qsiprep and stores them under the file structure described `above <#process-expected-output>`_.
+
+*DWIQC* comes with CUDA 10.1 pre-compiled inside the respective qsiprep and prequal containers to unburden users from compiling specific CUDA versions.
 
 **Prequal Output:**
 
@@ -360,7 +362,7 @@ To find the prequal pdf report, navigate to the ``--bids-dir`` directory you pas
 
 derivatives ---> dwiqc-prequal ---> subject_dir ---> session_dir ---> sub_session_dir_run__dwi ---> OUTPUTS ---> PDF ---> dtiQA.pdf
 
-Download an example :download:`here <examples/dtiQA.pdf>`.
+See `Prequal documentation <https://github.com/MASILab/PreQual?tab=readme-ov-file#outputs>`_ for more information.
 
 **Qsiprep Output:**
 
@@ -368,7 +370,7 @@ To find the qsiprep html report, navigate to the ``--bids-dir`` directory you pa
 
 derivatives ---> dwiqc-qsiprep ---> subject_dir ---> session_dir ---> sub_session_dir_run__dwi ---> qsiprep_output ---> qsiprep ---> sub-SUBJECT-imbedded_images.html
 
-Download an example :download:`here <examples/sub-MS881355-imbedded_images.html>`.
+See `Qsiprep documentation <https://qsiprep.readthedocs.io/en/latest/preprocessing.html>`_ for more information.
 
[…]
