### 2024-12-16 #6974 @danielasay
Hi @samhforbes and @jelleveraart I have responded to the issues raised above as detailed in this most recent commit to the github repository: https://github.com/harvard-nrg/dwiqc/commit/5b0202da1e00ec95767f05b68821be00d3f2dae3

Thank you for your feedback and support!
