+++ .github/workflows/pylint.yml
@@ -21,4 +21,4 @@ jobs:
         pip install .
     - name: Analysing the code with pylint
       run: |
-        pylint $(git ls-files '*.py') --errors-only --disable=E1101,E0606
+        pylint $(git ls-files '*.py') --errors-only --disable=E1101,E0606,E1121
