### 2025-01-06 #6974 @samhforbes
@danielasay can you set the licence on the archive to match the licence on GitHub please? And it looks as though the ordering of authors needs to be fixed.
