### 2025-01-06 #6974 @danielasay
@samhforbes Unfortunately, I couldn't find the exact same license as is listed on the Github repo so I chose the closest one that's still a BSD-3-Clause license. Sorry about the author order, I'll fix that.
