### 2025-01-06 #6974 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
