### 2025-01-09 #6974 @Kevin-Mattheus-Moerman
BSD-3-Clause license as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:

- [ ] Ensure you have a tagged release (listed here: https://github.com/harvard-nrg/dwiqc/tags) with a version tag that matches the one listed here and on Zenodo. We are using `2.0.0` here, so you could for instance create a version `2.0.0` to fix this.
- [ ] Ensure that the Zenodo  archive listed license matches your software license. You need to sign in and manually edit this.
- [ ] `tranparency` should read `transparency`
- [ ] `seemless` should read `seamless`
