### 2025-01-09 cbd3a9e
fix paper typos and update license
