+++ paper/paper.md
@@ -37,9 +37,9 @@ Diffusion weighted neuroimaging is an MR modality that maps white-matter microst
 
 # Statement of need
 
-Diffusion neuroimaging is a burgeoning field with huge potential to deepen our understanding of the brain. While exciting, it also means that acquisition parameters, study designs, and theoretical analysis frameworks vary greatly. `DWIQC` is an effort to make diffusion imaging analysis and quality assurance accessible to researchers with diverse experimental designs. By leveraging the respective strengths of various diffusion imaging analysis tools, `DWIQC` appeals to a broad range of users who employ different data acquisition approaches. Additionally, large, multi-site studies can can use the standardized preprocessing to ensure uniformity in data preprocessing, data quality metrics and downstream analysis. `DWIQC` serves as a one stop destination for users who want a seemless integration of Prequal, Qsiprep and FSL Eddy Quad with the flexibility of catering to diverse acquisition protocols. Users can configure settings in `DWIQC` that will be applied across all the software that is used, abstracting away the sometimes convoluted task of doing so for each software pipeline individually.
+Diffusion neuroimaging is a burgeoning field with huge potential to deepen our understanding of the brain. While exciting, it also means that acquisition parameters, study designs, and theoretical analysis frameworks vary greatly. `DWIQC` is an effort to make diffusion imaging analysis and quality assurance accessible to researchers with diverse experimental designs. By leveraging the respective strengths of various diffusion imaging analysis tools, `DWIQC` appeals to a broad range of users who employ different data acquisition approaches. Additionally, large, multi-site studies can can use the standardized preprocessing to ensure uniformity in data preprocessing, data quality metrics and downstream analysis. `DWIQC` serves as a one stop destination for users who want a seamless integration of Prequal, Qsiprep and FSL Eddy Quad with the flexibility of catering to diverse acquisition protocols. Users can configure settings in `DWIQC` that will be applied across all the software that is used, abstracting away the sometimes convoluted task of doing so for each software pipeline individually.
 
-Furthermore, `DWIQC's` integration with the XNAT data management platform facilitates ease of adoption due to XNAT's widespread use at neuroimaging centers. Sites that use `DWIQC` and upload its results to an XNAT instance provide tranparency in analysis practices for collaborating sites. Multi-site studies benefit by seeing diffusion analysis reports in the same format across sites and subjects.
+Furthermore, `DWIQC's` integration with the XNAT data management platform facilitates ease of adoption due to XNAT's widespread use at neuroimaging centers. Sites that use `DWIQC` and upload its results to an XNAT instance provide transparency in analysis practices for collaborating sites. Multi-site studies benefit by seeing diffusion analysis reports in the same format across sites and subjects.
 
 `DWIQC` benefits researchers by catching problems in data quality in real-time. The robustness `DWIQC's` pipelines allows even minor data quality issues to be caught. By running the diffusion weighted data through large-scale preprocessing pipelines, researchers are privy to data quality problems that would not be caught otherwise. As such, adjustments can be made mid-study to data collection protocols as necessary, rather than waiting until data collection has stopped only to discover data quality issues.
 
