### 2025-01-09 #6974 @danielasay
Hi @Kevin-Mattheus-Moerman, thanks for your comments. See my responses below:

1. There is now a version 2.0.0 tagged on github as well as version 2.0.0 uploaded to pypi.org
2. The License in the Github Repo and on Zenodo match as the BSD 3-Clause "New" or "Revised" License
3. Typo is fixed
4. Typo is fixed

Thank you for your support! Please let me know if there's anything else you need from me.
