### 2025-01-09 #6974 @Kevin-Mattheus-Moerman
@danielasay all looks good now. However, as the below image shows, the archive now lists two licenses which I think should only be your actual software license. We can proceed here now but I recommend you remove the creative commons license from the archive listing.

![image](https://github.com/user-attachments/assets/5609c0e8-4ff0-41ff-ae82-ba5bea13f9cc)
