### 2025-01-10 #6974 @Kevin-Mattheus-Moerman
@danielasay congratulations on this JOSS publication!!!!

@samhforbes thanks for editing!

And a special thank you to the reviewers: @mwegrzyn, @jelleveraart !!!!!!!!!!
