### 2025-01-10 #6974 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
