### 2025-01-10 #6974 @danielasay
@samhforbes @Kevin-Mattheus-Moerman I just came across a typo in the paper that I fixed, shown in this commit: https://github.com/harvard-nrg/dwiqc/commit/adaa0729ac685425cd8fd7c06115d21b480fce72

It's on the line that says, "Additionally, large, multi-site studies can use"

The old version had it written as, "studies can can use"

Is there a way to update the publication with the typo fix? Or is it set in stone? Thanks!
